commented out other tests and added DutchExchangeETHGNO.js test
